Update 35_bilinearNet_predictState_oneLevel.yml
Part of cleaning up what went really into the manuscript, here removing dependency on dataset with 600/0.05*1000 = 12.000.000 data-points, which was never used for any experiment that went in the manuscript except for here. Since training of the binlinear network initialized with 'analytic' parameters here was done with learning rate lr=0, the dataset size made absolutely no difference.